requirements.txt
# Web framework
fastapi==0.115.2

# ASGI server for FastAPI
uvicorn==0.23.2

# Data processing
pandas==2.2.2
numpy==2.2.5  # Required by pandas

# Dropbox API
dropbox==11.36.2

# Environment variable support
python-dotenv==1.0.1

# Date and time utilities
python-dateutil==2.9.0.post0
pytz==2025.2

# SQLite and database interaction
sqlalchemy==2.0.30

# FIT file parsing
fitparse==1.2.0

# HTTP requests
requests==2.32.3
